refactor: move import_frequency() to its own file
